requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2
django-axes==5.41.1
django-cors-headers==3.14.0
django-defender==0.9.7
django-dotenv==1.4.2
django-ipware==5.0.0
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.9
phonenumberslite==8.13.9
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.4
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.3
